+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## [2.9.1-next.5](https://github.com/warp-ds/elements/compare/v2.9.1-next.4...v2.9.1-next.5) (2026-05-08)
+
+
+### Bug Fixes
+
+* **button:** wrap react button wrapper and map camel case to snake case ([#646](https://github.com/warp-ds/elements/issues/646)) ([600a59c](https://github.com/warp-ds/elements/commit/600a59cca3f764a0743e6a58fe4d0a14125d53ba))
+
+## [2.9.1-next.4](https://github.com/warp-ds/elements/compare/v2.9.1-next.3...v2.9.1-next.4) (2026-05-06)
+
+
+### Bug Fixes
+
+* don't prevent default unless there's something to prevent ([97ba77a](https://github.com/warp-ds/elements/commit/97ba77a8199257e010410e0ae6bdd74aaf589818)), closes [#641](https://github.com/warp-ds/elements/issues/641)
+* fix a bug where all input would be stopped ([#644](https://github.com/warp-ds/elements/issues/644)) ([1468a03](https://github.com/warp-ds/elements/commit/1468a034b58fd5574af95a156117e20d21da09de))
+
 ## [2.9.1-next.3](https://github.com/warp-ds/elements/compare/v2.9.1-next.2...v2.9.1-next.3) (2026-05-04)